# server/go-kidsviewer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KidsViewer Backend API Documentation",
        "title": "KidsViewer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9988",
    "basePath": "/",
    "paths": {
...
        "/api/v1/watching/check": {
            "post": {
                "description": "Check the status of a watching session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watching"
                ],
                "summary": "Check watching session",
                "parameters": [
                    {
                        "description": "Check watching request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CheckWatchingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watching session status retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to check watching session",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/watching/history/{person_id}": {
            "get": {
                "description": "Get watching history for a specific person",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watching"
                ],
                "summary": "Get watching history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "person_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of sessions (default: 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watching history retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Person ID is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to get watching history",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/watching/skip": {
            "post": {
                "description": "Skip questions in a watching session (requires parental password)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watching"
                ],
                "summary": "Skip questions",
                "parameters": [
                    {
                        "description": "Skip questions request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.SkipQuestionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Questions skipped successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized - parental password required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to skip questions",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/watching/start": {
            "post": {
                "description": "Start a new watching session for a person",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watching"
                ],
                "summary": "Start watching session",
                "parameters": [
                    {
                        "description": "Start watching request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.StartWatchingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watching session started successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to start watching session",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/watching/verify": {
            "post": {
                "description": "Verify the answer to a question in a watching session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watching"
                ],
                "summary": "Verify question answer",
                "parameters": [
                    {
                        "description": "Question verification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.VerifyQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question verification completed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Watching session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to verify question",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.CheckWatchingRequest": {
            "type": "object",
            "required": [
                "watching_token"
            ],
            "properties": {
                "watching_token": {
                    "type": "string"
                }
            }
        },
        "services.SkipQuestionsRequest": {
            "type": "object",
            "required": [
                "password",
                "watching_token"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "watching_token": {
                    "type": "string"
                }
            }
        },
        "services.StartWatchingRequest": {
            "type": "object",
            "required": [
                "person_id",
                "platform_url"
            ],
            "properties": {
                "person_id": {
                    "type": "string"
                },
                "platform_url": {
                    "type": "string"
                }
            }
        },
        "services.VerifyQuestionRequest": {
            "type": "object",
            "required": [
                "answer",
                "question_id",
                "watching_token"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                },
                "watching_token": {
                    "type": "string"
                }
            }
        }
    }
}